% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_map_to_sf.R
\name{extract_spatial_data}
\alias{extract_spatial_data}
\title{extract the spatial data from the parsed results of a \code{time_map} call}
\usage{
extract_spatial_data(result)
}
\arguments{
\item{result}{a single row from list of results}
}
\description{
extract the spatial data from the parsed results of a \code{time_map} call
}
