% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multi_search}
\alias{multi_search}
\title{Multiple pulled and/or merged isochrones}
\format{
An object of class \code{traveltime_api} of length 2.
}
\source{
\url{https://github.com/traveltime-dev/traveltime-sdk-r}
}
\usage{
multi_search
}
\description{
Multiple pulled and/or merged isochrones
}
\keyword{datasets}
