% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{single_search}
\alias{single_search}
\title{Traveltime's default example of pulling isochrones}
\format{
An object of class \code{traveltime_api} of length 2.
}
\source{
\url{https://github.com/traveltime-dev/traveltime-sdk-r}
}
\usage{
single_search
}
\description{
Traveltime's default example of pulling isochrones
}
\keyword{datasets}
