% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_map_sf.R
\name{time_map_sf}
\alias{time_map_sf}
\title{convenience wrapper that combines time_map and time_map_to_sf}
\usage{
time_map_sf(return_list = FALSE, ...)
}
\arguments{
\item{return_list}{if FALSE (default), returns all results as a single
\code{sf} dataframe.}

\item{...}{arguments passed to \code{time_map}}
}
\description{
convenience wrapper that combines time_map and time_map_to_sf
}
